Move ServiceClient interfaces to driver-definitions, reducing layerInfo diff
Interfaces implemented by drivers and used by the runtime conventionally
belong in @fluidframework/driver-definitions (same pattern as
IDocumentServiceFactory, IDocumentService, etc.).

Moves Registry, RegistryKey, basicKey, registryLookup, ServiceClient,
FluidContainer, DataStoreKind, DataStoreKey, DataStoreRegistry,
FluidContainerWithService, and FluidContainerAttached from
runtime-definitions to driver-definitions. Moves serviceClientUtils.ts
from driver-utils to runtime-definitions (using IContainerRuntimeBase
instead of IContainerRuntime to avoid a circular dependency).

runtime-definitions re-exports all moved types so existing consumers
require no import path changes.

This eliminates three layerInfo additions introduced by the branch:
the Runtime-Definitions Definitions layer group, the Driver-Utils →
Runtime-Definitions dep, and the Runtime → Runtime-Definitions dep.
The remaining layerInfo diff (Tinylicious-Driver split) is unavoidable
because ContainerRuntime is used directly.
